requirements.txt
fastapi
uvicorn[standard]
pymongo
argon2-cffi
python-multipart